refactor: #3 日時のフォーマット処理をカスタムパイプとして定義。
