+++ bear-clone/src/app/note-list/note-item/note-item.component.html
@@ -8,5 +8,5 @@
 >
   <h1 class="text-sm font-bold">{{ note.title }}</h1>
   <p class="line-clamp-2 text-sm text-[#888889]">{{ note.content }}</p>
-  <span class="text-xs text-[#888889]">{{ formatUpdateAt(note.updatedAt) }}</span>
+  <span class="text-xs text-[#888889]">{{ note.updatedAt | formatDate }}</span>
 </div>
