+++ bear-clone/src/app/note-list/note-item/note-item.component.spec.ts
@@ -24,13 +24,5 @@ describe("NoteItemComponent", () => {
 
   it("メモ本文が表示されている", () => {});
 
-  describe("formatUpdateAt", () => {
-    it("メモ作成から1分以内なら『~~ seconds ago』と表示される。", () => {});
-
-    it("メモ作成から1分以上1時間以内なら『~~ minutes ago』と表示される。", () => {});
-
-    it("メモ作成から1時間以上1日以内なら『~~ hours ago』と表示される。", () => {});
-
-    it("メモ作成から1日以上なら『Jan 26』の様に表示される。", () => {});
-  });
+  it("更新日時が表示されている。", () => {});
 });
